+++ src/common/messaging/get-packages.ts
@@ -1,9 +1,11 @@
 type GetPackagesRequest = {
   Url: string;
+  SourceName?: string;
   Filter: string;
   Prerelease: boolean;
   Skip: number;
   Take: number;
+  PasswordScriptPath?: string;
 };
 
 type GetPackagesResponse = {